git-svn-id: file:///semplestb/final_repo@550 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMplestBidding/src/semplest/bidding/estimation/EstimatorData.java
@@ -24,12 +24,13 @@ public class EstimatorData {
 		return targetVsBid.get(bid);
 	}
 	
-	public Double [] getBidArray(){
-		Double [] bidArray = new Double[targetVsBid.size()];
+	
+	public double [] getBidArray(){
+		double [] bidArray = new double[targetVsBid.size()];
 		Set<Double> bidSet = targetVsBid.keySet();
 		int i=0;
 		for(Double bid : bidSet){
-			bidArray[i]=bid;
+			bidArray[i]=bid.doubleValue();
 			i++;
 		}
 		return bidArray;
